Split-up admin make model test
